+++ packages/account-api/src/api/provider.ts
@@ -5,6 +5,13 @@ import type { KeyringAccount } from '@metamask/keyring-api';
  */
 export type AccountProvider<Account extends KeyringAccount> = {
   /**
+   * Gets an account for a given ID.
+   *
+   * @returns An account, or undefined if not found.
+   */
+  getAccount: (id: Account['id']) => Account | undefined;
+
+  /**
    * Gets all accounts for a given entropy source and group index.
    *
    * @returns A list of all account for this provider.